Added support for RTMP (Flash Media Server).  In files.xml, point a file directly at a fms server/path (i.e.)  rtmp://SOMESERVER/SOMEAPP/someflv.flv.  See ProtocolRTMP
ControlPlugin now creates a plugin automatically (see ContentMP3 -- and how it can automatically have a visualizer).

Controls now support a child control.  All controls of the parent and child will be visible.  I did this to try to get a generic tempo effect, that uses any bitmap filter.  Not done, but the pieces are there.

Added beginning remnants of a filter scroller.
Added beginning remnants of a tempo filter that randomly adjusts framerate (trying to make it look cooler -- but so far it's ok)

Various Bug Fixes